TEST/DJReact/requirements.txt
djangorestframework
markdown
django-filter
Django==4.1
asgirefdjango-cors-headers
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
PyJWT
pytz
sqlparse